+++ e2e-tests/development-runtime/src/pages/error-handling/query-validation-error.js
@@ -1,0 +1,19 @@
+import React from "react"
+import { graphql } from "gatsby"
+
+function QueryValidationError() {
+  return <p data-testid="hot">Working</p>
+}
+
+export default QueryValidationError
+
+export const query = graphql`
+  {
+    site {
+      siteMetadata {
+        title
+        # query-validation-error
+      }
+    }
+  }
+`