+++ packages/gatsby-plugin-image/src/node-apis/node-utils.ts
@@ -1,0 +1,4 @@
+import path from "path"
+
+export const getCacheDir = (root: string): string =>
+  path.join(root, `.cache`, `caches`, `gatsby-plugin-image`)