+++ packages/gatsby-plugin-utils/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./validate"
+export * from "./test-plugin-options-schema"
+export * from "./joi"