+++ packages/gatsby-source-contentful/src/gatsby-node.js
@@ -2,10 +2,18 @@ const path = require(`path`)
 const isOnline = require(`is-online`)
 const _ = require(`lodash`)
 const fs = require(`fs-extra`)
+const { createClient } = require(`contentful`)
+const v8 = require(`v8`)
+const fetch = require(`node-fetch`)
+const { Joi } = require(`gatsby-plugin-utils`)
 
 const normalize = require(`./normalize`)
 const fetchData = require(`./fetch`)
-const { createPluginConfig, validateOptions } = require(`./plugin-options`)
+const {
+  createPluginConfig,
+  maskText,
+  formatPluginOptionsForCLI,
+} = require(`./plugin-options`)
 const { downloadContentfulAssets } = require(`./download-contentful-assets`)
 
 const conflictFieldPrefix = `contentful`
@@ -22,7 +30,144 @@ const restrictedNodeFields = [
 
 exports.setFieldsOnGraphQLNodeType = require(`./extend-node-type`).extendNodeType
 
-exports.onPreBootstrap = validateOptions
+// TODO: Remove once pluginOptionsSchema is stable
+exports.onPreInit = ({ reporter }, options) => {
+  const result = pluginOptionsSchema({ Joi }).validate(options, {
+    abortEarly: false,
+    externals: false,
+  })
+  if (result.error) {
+    const errors = {}
+    result.error.details.forEach(detail => {
+      errors[detail.path[0]] = detail.message
+    })
+    reporter.panic(`Problems with gatsby-source-contentful plugin options:
+${formatPluginOptionsForCLI(options, errors)}`)
+  }
+
+  options = result.value
+}
+
+const validateContentfulAccess = async pluginOptions => {
+  if (process.env.NODE_ENV === `test`) return undefined
+
+  await fetch(`https://${pluginOptions.host}/spaces/${pluginOptions.spaceId}`, {
+    headers: {
+      Authorization: `Bearer ${pluginOptions.accessToken}`,
+      "Content-Type": `application/json`,
+    },
+  })
+    .then(res => res.ok)
+    .then(ok => {
+      if (!ok)
+        throw new Error(
+          `Cannot access Contentful space "${maskText(
+            pluginOptions.spaceId
+          )}" with access token "${maskText(
+            pluginOptions.accessToken
+          )}". Make sure to double check them!`
+        )
+    })
+
+  return undefined
+}
+
+const pluginOptionsSchema = ({ Joi }) =>
+  Joi.object()
+    .keys({
+      accessToken: Joi.string()
+        .description(
+          `Contentful delivery api key, when using the Preview API use your Preview API key`
+        )
+        .required()
+        .empty(),
+      spaceId: Joi.string()
+        .description(`Contentful spaceId`)
+        .required()
+        .empty(),
+      host: Joi.string()
+        .description(
+          `The base host for all the API requests, by default it's 'cdn.contentful.com', if you want to use the Preview API set it to 'preview.contentful.com'. You can use your own host for debugging/testing purposes as long as you respect the same Contentful JSON structure.`
+        )
+        .default(`cdn.contentful.com`)
+        .empty(),
+      environment: Joi.string()
+        .description(
+          `The environment to pull the content from, for more info on environments check out this [Guide](https://www.contentful.com/developers/docs/concepts/multiple-environments/).`
+        )
+        .default(`master`)
+        .empty(),
+      downloadLocal: Joi.boolean()
+        .description(
+          `Downloads and caches ContentfulAsset's to the local filesystem. Allows you to query a ContentfulAsset's localFile field, which is not linked to Contentful's CDN. Useful for reducing data usage.
+You can pass in any other options available in the [contentful.js SDK](https://github.com/contentful/contentful.js#configuration).`
+        )
+        .default(false),
+      localeFilter: Joi.func()
+        .description(
+          `Possibility to limit how many locales/nodes are created in GraphQL. This can limit the memory usage by reducing the amount of nodes created. Useful if you have a large space in contentful and only want to get the data from one selected locale.
+For example, to filter locales on only germany \`localeFilter: locale => locale.code === 'de-DE'\`
+
+List of locales and their codes can be found in Contentful app -> Settings -> Locales`
+        )
+        .default(() => true),
+      forceFullSync: Joi.boolean()
+        .description(
+          `Prevents the use of sync tokens when accessing the Contentful API.`
+        )
+        .default(false),
+      pageLimit: Joi.number()
+        .integer()
+        .description(
+          `Number of entries to retrieve from Contentful at a time. Due to some technical limitations, the response payload should not be greater than 7MB when pulling content from Contentful. If you encounter this issue you can set this param to a lower number than 100, e.g 50.`
+        )
+        .default(100),
+      assetDownloadWorkers: Joi.number()
+        .integer()
+        .description(
+          `Number of workers to use when downloading contentful assets. Due to technical limitations, opening too many concurrent requests can cause stalled downloads. If you encounter this issue you can set this param to a lower number than 50, e.g 25.`
+        )
+        .default(50),
+      proxy: Joi.object()
+        .keys({
+          host: Joi.string().required(),
+          port: Joi.number().required(),
+          auth: Joi.object().keys({
+            username: Joi.string(),
+            password: Joi.string(),
+          }),
+        })
+        .description(
+          `Axios proxy configuration. See the [axios request config documentation](https://github.com/mzabriskie/axios#request-config) for further information about the supported values.`
+        ),
+      useNameForId: Joi.boolean()
+        .description(
+          `Use the content's \`name\` when generating the GraphQL schema e.g. a Content Type called \`[Component] Navigation bar\` will be named \`contentfulComponentNavigationBar\`.
+    When set to \`false\`, the content's internal ID will be used instead e.g. a Content Type with the ID \`navigationBar\` will be called \`contentfulNavigationBar\`.
+
+    Using the ID is a much more stable property to work with as it will change less often. However, in some scenarios, Content Types' IDs will be auto-generated (e.g. when creating a new Content Type without specifying an ID) which means the name in the GraphQL schema will be something like \`contentfulC6XwpTaSiiI2Ak2Ww0oi6qa\`. This won't change and will still function perfectly as a valid field name but it is obviously pretty ugly to work with.
+
+    If you are confident your Content Types will have natural-language IDs (e.g. \`blogPost\`), then you should set this option to \`false\`. If you are unable to ensure this, then you should leave this option set to \`true\` (the default).`
+        )
+        .default(true),
+      // default plugins passed by gatsby
+      plugins: Joi.array(),
+      richText: Joi.object()
+        .keys({
+          resolveFieldLocales: Joi.boolean()
+            .description(
+              `If you want to resolve the locales in fields of assets and entries that are referenced by rich text (e.g., via embedded entries or entry hyperlinks), set this to \`true\`. Otherwise, fields of referenced assets or entries will be objects keyed by locale.`
+            )
+            .default(false),
+        })
+        .default({}),
+    })
+    .external(validateContentfulAccess)
+
+if (process.env.GATSBY_EXPERIMENTAL_PLUGIN_OPTION_VALIDATION) {
+  exports.pluginOptionsSchema = pluginOptionsSchema
+}
+
 /***
  * Localization algorithm
  *
@@ -45,75 +190,194 @@ exports.sourceNodes = async (
     cache,
     getCache,
     reporter,
+    schema,
+    parentSpan,
   },
   pluginOptions
 ) => {
-  const { createNode, deleteNode, touchNode, setPluginStatus } = actions
+  const { createNode, deleteNode, touchNode } = actions
 
-  const online = await isOnline()
-
-  // If the user knows they are offline, serve them cached result
-  // For prod builds though always fail if we can't get the latest data
-  if (
-    !online &&
-    process.env.GATSBY_CONTENTFUL_OFFLINE === `true` &&
-    process.env.NODE_ENV !== `production`
-  ) {
-    getNodes().forEach(node => {
-      if (node.internal.owner !== `gatsby-source-contentful`) {
-        return
-      }
-      touchNode({ nodeId: node.id })
-      if (node.localFile___NODE) {
-        // Prevent GraphQL type inference from crashing on this property
-        touchNode({ nodeId: node.localFile___NODE })
-      }
-    })
-
-    reporter.info(`Using Contentful Offline cache ⚠️`)
+  let currentSyncData, contentTypeItems, defaultLocale, locales, space
+  if (process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE) {
     reporter.info(
-      `Cache may be invalidated if you edit package.json, gatsby-node.js or gatsby-config.js files`
+      `GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE: Storing/loading remote data through \`` +
+        process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE +
+        `\` so Remote changes CAN NOT be detected!`
     )
-
-    return
   }
+  const forceCache = await fs.exists(
+    process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE
+  )
 
   const pluginConfig = createPluginConfig(pluginOptions)
+  const sourceId = `${pluginConfig.get(`spaceId`)}-${pluginConfig.get(
+    `environment`
+  )}`
 
-  const createSyncToken = () =>
-    `${pluginConfig.get(`spaceId`)}-${pluginConfig.get(
-      `environment`
-    )}-${pluginConfig.get(`host`)}`
+  const CACHE_SYNC_TOKEN = `contentful-sync-token-${sourceId}`
+  const CACHE_SYNC_DATA = `contentful-sync-data-${sourceId}`
 
-  // Get sync token if it exists.
-  let syncToken
-  if (
-    !pluginConfig.get(`forceFullSync`) &&
-    store.getState().status.plugins &&
-    store.getState().status.plugins[`gatsby-source-contentful`] &&
-    store.getState().status.plugins[`gatsby-source-contentful`][
-      createSyncToken()
-    ]
-  ) {
-    syncToken = store.getState().status.plugins[`gatsby-source-contentful`][
-      createSyncToken()
-    ]
+  /*
+   * Subsequent calls of Contentfuls sync API return only changed data.
+   *
+   * In some cases, especially when using rich-text fields, there can be data
+   * missing from referenced entries. This breaks the reference matching.
+   *
+   * To workround this, we cache the initial sync data and merge it
+   * with all data from subsequent syncs. Afterwards the references get
+   * resolved via the Contentful JS SDK.
+   */
+  let syncToken = await cache.get(CACHE_SYNC_TOKEN)
+  let previousSyncData = {
+    assets: [],
+    entries: [],
+  }
+  let cachedData = await cache.get(CACHE_SYNC_DATA)
+
+  if (cachedData) {
+    previousSyncData = cachedData
   }
 
-  const {
-    currentSyncData,
-    contentTypeItems,
-    defaultLocale,
-    locales,
-    space,
-  } = await fetchData({
-    syncToken,
-    reporter,
-    pluginConfig,
+  if (forceCache) {
+    // If the cache has data, use it. Otherwise do a remote fetch anyways and prime the cache now.
+    // If present, do NOT contact contentful, skip the round trips entirely
+    reporter.info(
+      `GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE was set. Skipping remote fetch, using data stored in`,
+      process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE
+    )
+    ;({
+      currentSyncData,
+      contentTypeItems,
+      defaultLocale,
+      locales,
+      space,
+    } = v8.deserialize(
+      fs.readFileSync(process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE)
+    ))
+  } else {
+    const online = await isOnline()
+
+    // If the user knows they are offline, serve them cached result
+    // For prod builds though always fail if we can't get the latest data
+    if (
+      !online &&
+      process.env.GATSBY_CONTENTFUL_OFFLINE === `true` &&
+      process.env.NODE_ENV !== `production`
+    ) {
+      getNodes().forEach(node => {
+        if (node.internal.owner !== `gatsby-source-contentful`) {
+          return
+        }
+        touchNode({ nodeId: node.id })
+        if (node.localFile___NODE) {
+          // Prevent GraphQL type inference from crashing on this property
+          touchNode({ nodeId: node.localFile___NODE })
+        }
+      })
+
+      reporter.info(`Using Contentful Offline cache ⚠️`)
+      reporter.info(
+        `Cache may be invalidated if you edit package.json, gatsby-node.js or gatsby-config.js files`
+      )
+
+      return
+    }
+    if (process.env.GATSBY_CONTENTFUL_OFFLINE) {
+      reporter.info(
+        `Note: \`GATSBY_CONTENTFUL_OFFLINE\` was set but it either was not \`true\`, we _are_ online, or we are in production mode, so the flag is ignored.`
+      )
+    }
+
+    const fetchActivity = reporter.activityTimer(
+      `Contentful: Fetch data (${sourceId})`,
+      {
+        parentSpan,
+      }
+    )
+
+    fetchActivity.start()
+    ;({
+      currentSyncData,
+      contentTypeItems,
+      defaultLocale,
+      locales,
+      space,
+    } = await fetchData({
+      syncToken,
+      reporter,
+      pluginConfig,
+      parentSpan,
+    }))
+
+    if (process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE) {
+      reporter.info(
+        `GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE was set. Writing v8 serialized glob of remote data to: ` +
+          process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE
+      )
+      fs.writeFileSync(
+        process.env.GATSBY_CONTENTFUL_EXPERIMENTAL_FORCE_CACHE,
+        v8.serialize({
+          currentSyncData,
+          contentTypeItems,
+          defaultLocale,
+          locales,
+          space,
+        })
+      )
+    }
+    fetchActivity.end()
+  }
+
+  const processingActivity = reporter.activityTimer(
+    `Contentful: Proccess data (${sourceId})`,
+    {
+      parentSpan,
+    }
+  )
+  processingActivity.start()
+
+  // Create a map of up to date entries and assets
+  function mergeSyncData(previous, current, deleted) {
+    const entryMap = new Map()
+    previous.forEach(e => !deleted.has(e.sys.id) && entryMap.set(e.sys.id, e))
+    current.forEach(e => !deleted.has(e.sys.id) && entryMap.set(e.sys.id, e))
+    return [...entryMap.values()]
+  }
+
+  const deletedSet = new Set(currentSyncData.deletedEntries.map(e => e.sys.id))
+
+  const mergedSyncData = {
+    entries: mergeSyncData(
+      previousSyncData.entries,
+      currentSyncData.entries,
+      deletedSet
+    ),
+    assets: mergeSyncData(
+      previousSyncData.assets,
+      currentSyncData.assets,
+      deletedSet
+    ),
+  }
+
+  // Store a raw and unresolved copy of the data for caching
+  const mergedSyncDataRaw = _.cloneDeep(mergedSyncData)
+
+  // Use the JS-SDK to resolve the entries and assets
+  const res = createClient({
+    space: `none`,
+    accessToken: `fake-access-token`,
+  }).parseEntries({
+    items: mergedSyncData.entries,
+    includes: {
+      assets: mergedSyncData.assets,
+      entries: mergedSyncData.entries,
+    },
   })
 
+  mergedSyncData.entries = res.items
+
   const entryList = normalize.buildEntryList({
-    currentSyncData,
+    mergedSyncData,
     contentTypeItems,
   })
 
@@ -122,12 +386,17 @@ exports.sourceNodes = async (
   // are "updated" so will get the now deleted reference removed.
 
   function deleteContentfulNode(node) {
+    const normalizedType = node.sys.type.startsWith(`Deleted`)
+      ? node.sys.type.substring(`Deleted`.length)
+      : node.sys.type
+
     const localizedNodes = locales
       .map(locale => {
         const nodeId = createNodeId(
           normalize.makeId({
             spaceId: space.sys.id,
             id: node.sys.id,
+            type: normalizedType,
             currentLocale: locale.code,
             defaultLocale,
           })
@@ -151,21 +420,22 @@ exports.sourceNodes = async (
   )
   existingNodes.forEach(n => touchNode({ nodeId: n.id }))
 
-  const assets = currentSyncData.assets
+  const assets = mergedSyncData.assets
 
   reporter.info(`Updated entries ${currentSyncData.entries.length}`)
   reporter.info(`Deleted entries ${currentSyncData.deletedEntries.length}`)
   reporter.info(`Updated assets ${currentSyncData.assets.length}`)
   reporter.info(`Deleted assets ${currentSyncData.deletedAssets.length}`)
-  console.timeEnd(`Fetch Contentful data`)
 
   // Update syncToken
   const nextSyncToken = currentSyncData.nextSyncToken
 
-  // Store our sync state for the next sync.
-  const newState = {}
-  newState[createSyncToken()] = nextSyncToken
-  setPluginStatus(newState)
+  await Promise.all([
+    cache.set(CACHE_SYNC_DATA, mergedSyncDataRaw),
+    cache.set(CACHE_SYNC_TOKEN, nextSyncToken),
+  ])
+
+  reporter.verbose(`Building Contentful reference map`)
 
   // Create map of resolvable ids so we can check links against them while creating
   // links.
@@ -189,36 +459,60 @@ exports.sourceNodes = async (
     useNameForId: pluginConfig.get(`useNameForId`),
   })
 
+  reporter.verbose(`Resolving Contentful references`)
+
   const newOrUpdatedEntries = []
   entryList.forEach(entries => {
     entries.forEach(entry => {
-      newOrUpdatedEntries.push(entry.sys.id)
+      newOrUpdatedEntries.push(`${entry.sys.id}___${entry.sys.type}`)
     })
   })
 
   // Update existing entry nodes that weren't updated but that need reverse
   // links added.
   existingNodes
-    .filter(n => _.includes(newOrUpdatedEntries, n.id))
+    .filter(n => _.includes(newOrUpdatedEntries, `${n.id}___${n.sys.type}`))
     .forEach(n => {
-      if (foreignReferenceMap[n.id]) {
-        foreignReferenceMap[n.id].forEach(foreignReference => {
-          // Add reverse links
-          if (n[foreignReference.name]) {
-            n[foreignReference.name].push(foreignReference.id)
-            // It might already be there so we'll uniquify after pushing.
-            n[foreignReference.name] = _.uniq(n[foreignReference.name])
-          } else {
-            // If is one foreign reference, there can always be many.
-            // Best to be safe and put it in an array to start with.
-            n[foreignReference.name] = [foreignReference.id]
+      if (foreignReferenceMap[`${n.id}___${n.sys.type}`]) {
+        foreignReferenceMap[`${n.id}___${n.sys.type}`].forEach(
+          foreignReference => {
+            // Add reverse links
+            if (n[foreignReference.name]) {
+              n[foreignReference.name].push(foreignReference.id)
+              // It might already be there so we'll uniquify after pushing.
+              n[foreignReference.name] = _.uniq(n[foreignReference.name])
+            } else {
+              // If is one foreign reference, there can always be many.
+              // Best to be safe and put it in an array to start with.
+              n[foreignReference.name] = [foreignReference.id]
+            }
           }
-        })
+        )
       }
     })
 
+  processingActivity.end()
+
+  const creationActivity = reporter.activityTimer(
+    `Contentful: Create nodes (${sourceId})`,
+    {
+      parentSpan,
+    }
+  )
+  creationActivity.start()
+
   for (let i = 0; i < contentTypeItems.length; i++) {
     const contentTypeItem = contentTypeItems[i]
+
+    if (entryList[i].length) {
+      reporter.info(
+        `Creating ${entryList[i].length} Contentful ${
+          pluginConfig.get(`useNameForId`)
+            ? contentTypeItem.name
+            : contentTypeItem.sys.id
+        } nodes`
+      )
+    }
 
     // A contentType can hold lots of entries which create nodes
     // We wait until all nodes are created and processed until we handle the next one
@@ -243,6 +537,10 @@ exports.sourceNodes = async (
     )
   }
 
+  if (assets.length) {
+    reporter.info(`Creating ${assets.length} Contentful asset nodes`)
+  }
+
   for (let i = 0; i < assets.length; i++) {
     // We wait for each asset to be process until handling the next one.
     await Promise.all(
@@ -257,7 +555,11 @@ exports.sourceNodes = async (
     )
   }
 
+  creationActivity.end()
+
   if (pluginConfig.get(`downloadLocal`)) {
+    reporter.info(`Download Contentful asset files`)
+
     await downloadContentfulAssets({
       actions,
       createNodeId,
@@ -266,6 +568,7 @@ exports.sourceNodes = async (
       getCache,
       getNodesByType,
       reporter,
+      assetDownloadWorkers: pluginConfig.get(`assetDownloadWorkers`),
     })
   }