+++ packages/gatsby-plugin-stylus/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-stylus@2.3.12...gatsby-plugin-stylus@2.3.13) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-plugin-stylus
+
 ## [2.3.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-stylus@2.3.11...gatsby-plugin-stylus@2.3.12) (2020-09-28)
 
 ### Bug Fixes